requirements.txt
fastapi
uvicorn
google-generativeai
youtube-search
pydantic
python-multipart
python-dotenv